Refactor Navbar & added PropTypes
